services/june-auth/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
pyotp==2.9.0
qrcode[pil]==7.4.2
cryptography==43.0.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-dotenv==1.0.1
